requirements.txt
polars
pandas
numpy
scikit-learn
plotly
seaborn
matplotlib
requests
pyarrow
pyyaml
streamlit
fastapi
uvicorn
shap
matplotlib
optuna
mlflow
pytest
httpx
psycopg2-binary